+++ src/Extractors/Extractor.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Sowren\SvgAvatarGenerator\Extractors;
+
+interface Extractor
+{
+    /**
+     * Extract initials from given text.
+     */
+    public function extract(string $text): string;
+}